refactor DrawStep interfaces into its own file
removes a circular dependency
